Refactor: Move JWT classes to service package
